Persistencia de Picture.kt
Se han modificado los archivos necesarios para incluir el tipo de imagen por si es de Resource o Internalfile.
Se ha creado la persistencia para incluir las Pictures en la DDBB además se han creado dos métodos que devuelven: un listado de pictures no usados o una picture no usado

+++ 4puzzleGame/app/src/main/java/com/puzzle/game/adapters/ScoreListAdapter.kt
@@ -57,7 +57,7 @@ class ScoreListAdapter(
         fecha.text = bestScore.fechaFin.toString()
         puntuacion.text = bestScore.score.toString()
         println("bestScore.idImagen = ${bestScore.idImagen}")
-        imagenScore.setImageResource(bestScore.idImagen);
+        imagenScore.setImageResource(bestScore.idImagen.toInt());
 
 // Get detail element
